requirements.txt
annotated-types==0.7.0
anyio==4.10.0
APScheduler==3.6.3
cachetools==4.2.2
certifi==2025.8.3
charset-normalizer==3.4.3
colorama==0.4.6
distro==1.9.0
dotenv-python==0.0.1
exceptiongroup==1.3.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
jiter==0.10.0
openai==1.100.1
pydantic==2.11.7
pydantic_core==2.33.2
python-telegram-bot==13.15
pytz==2025.2
requests==2.32.5
six==1.17.0
sniffio==1.3.1
tinydb==4.8.2
tornado==6.1
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
tzlocal==5.3.1
urllib3==2.5.0
